+++ src/fileio/constants.ts
@@ -1,0 +1,10 @@
+import type { MessageOfparsePluginParamRecordEx } from "./types";
+
+export const READ_PLUGIN_MESSAGES: MessageOfparsePluginParamRecordEx = {
+  readErrorPluginsJS: "Failed to read plugins.js",
+  readErrorPluginBody: "Failed to read plugin file",
+  parseError: "Failed to parse plugin",
+  notArray: "Plugin format is invalid: not an array",
+  partialSuccess: "Some plugins failed to read or parse",
+  success: "All plugins read and parsed successfully",
+};
